Changing unicode character to indicate urgency which resolves #1388 + fixing Eisenhower Matrices shortcut

+++ app/src/qt/widgets/urgency_combo_box.h
@@ -36,7 +36,7 @@ public:
     ~UrgencyComboBox();
 
     int8_t getValue() {
-        return itemData(currentIndex(), Qt::UserRole).value<int>();
+        return static_cast<int8_t>(itemData(currentIndex(), Qt::UserRole).value<int>());
     }
 };
